Fix for #1766
Not sure I've added this in the right place, or if there are better ways to check the DB connection type, but it fixes an issue that occurs when rolling back the migrations containing enum types on MySQL connections. I have no clue if this could have any side effects, but it seems to work fine.