Improved datepicker selection
Changed datepicker to: https://nehakadam.github.io/DateTimePicker/

It's responsive, and allows for user-input. Which the current one doesn't as far as I can see, and that's a hassle since the current one defaults to 1970, and going from 1970 to preset time takes a while, when you can only go forwards one year at a time.

It's also quite stylish.