Address .login-sidebar properly for CSS overrule
Currently, when the primary color in the config is changed, it is not reflected on the border on the .login-sidebar div. The reason for this is that it isn't properly address and thus can't be overruled. This PR proposes addressing the sidebar so that it will overrule the default setting and color.