PHP7 compatibility fix for spatial traits
Getting lat and lng data from point value with the list function is ambiguous and incoherent between PHP5 and PHP7
This lead to wrong or right location coords depending on verion of php used
Excerpt from PHP.net documentation below (http://php.net/manual/en/function.list.php)

Warning
In PHP 5, list() assigns the values starting with the right-most parameter. In PHP 7, list() starts with the left-most parameter.

If you are using plain variables, you don't have to worry about this. But if you are using arrays with indices you usually expect the order of the indices in the array the same you wrote in the list() from left to right, which is not the case in PHP 5, as it's assigned in the reverse order.

Generally speaking, it is advisable to avoid relying on a specific order of operation, as this may change again in the future.