Fix column not found when using additional_attributes w/ xMany relations
Fix for belongsToMany and hasMany relationship display with
additional_attributes. You can now use additional_attributes with those
relations.
#2885

The fix get the model before plucking it. It caused error because when
you pluck the query builder, it affects the mysql query. When you use
additional attribute, it doesn’t exist in the database and result with
a `SQLSTATE[42S22]: Column not found: 1054 Unknown column ‘youradditionalattribute’ in 'field list’` error…