Add double click function to BREAD row
Add a way the user can navigate through the BREAD views. By double
clicking the row, the user goes to the view or edit page, depending
on their rights. To simplify the way to delete a single item, add a
delete button in the browse view if the user has rights to delete it.