Update relationship.blade.php
Relationship formfield with ajax data load and infinite scrolling. It loads 20 records at a time, it adds 20 more records for each scroll, it's searchable, it shows the preselected values, it works with belongsTo and belongsToMany, it works with multiple relationship formfields on the same view. Note: we could add a json option in the bread editor to change the number of records per page.