# Hidden ID Field Fix, Attempt Deux #
If I understand the underlying cause correctly, the default `id` field is being saved with a `belongsToMany` relationship, which makes it appear as though the `id` field is connected to it, when it shouldn't be. Thus, we stop it from making such a 'connection' by only saving the field if if the relationship is a `belongsTo` relationship.

* Restores the commented-out lines that were clearly not a part of the problem.
* Adds a check to see if a given relationship is `belongsTo` before hiding any fields associated with it.