Proposed fix for multiple image field quality setting
This is a duplicate PR of https://github.com/the-control-group/voyager/pull/3507 on the 1.1 branch as requested.

The multiple image content type class ignores user defined quality settings.

The class does an isset() check on the $options variable, which is not defined thus always returns false. As a result the $resize_quality variable is set to it's default of 75%.

This PR fixes the conditional check.