First parameter must either be an object or the name of an existing class
First parameter must either be an object or the name of an existing class issue arises when in create timestamp field in any entity. Id we do not specify any details options in bread, it stores in db null, so at the time of browse, it gives error.
So this is my workaround.