Fix the route problem after adding/updating BREAD.
When we add new BREAD the new slug appends to route list. In new Laravel version we have to manually cache the list of routes to have it updated after modifications.

In our case, when we add new BREAD and it's routes append to the route list we have to cache the list again. Because of it we have an exception every time we create new BREAD (update old one slug).

For two cases (store and update) I call here Artisan command to optimize the config files.