Feature: Added Modals, Added ability to add full detail of a model when has relationships
Voyager by default when a model have a relationship it adds only the field that's been added when making the bread. But this PR adds the ability to show a button beside those fields that have relationship and when clicked showing the complete form of that related bread. It's been tested all working well when adding filling the form it saves the data without leaving the page. what is missing is a javascript or ajax function to update the current row that have the button beside.