Show checkboxes column if mass action is displayed
bread.browse didn't know if mass action was going to be displayed or not because check was done inside bread.partials.actions and I don't think there is an easy way to pass data back from an @include also policy for action (not mass action) might depend on record data so we can't check once for all.

What I did was to move shouldActionDisplayOnDataType check to the controller VoyagerBaseController that is now passing to view an array of actions objects, in bread.partials.actions stays the policy check that might depend on record data.