Fix custom formfield view not aligned with default views
Passing $options to custom views makes them more in line with default ones.
Also for some reason in custom views `'action' => 'browse'` is used instead of `'view' => 'browse'` so I added the latter and left the former for compatibility reasons, I'd suggest we deprecate `action` so we can remove it in a future release.

Closes #4626