Let user select key for all relationships
Right now the key to store is selectable only for BelongsTo relationships.
For other types the property `key` is still saved taking the first column of the table, but it can't be chosen.

This PR also adds support for more advanced parameters in BelongsToMany relationships, there is no interface yet so they need to be set in the details field, once implemented within the graphic interface they'll simply disappear from details field like the other relationship parameters.

Closes #2648