Fix prepare translations check
Right all traslatable fields need to be sent for add/update to succed.

This PR intersect translatable with rows enabled for add/edit, so that check is performed only for rows that are supposed to be saved.

While saving BREAD configuration this intersecttion is ignored.

There is another PR #4620 fixing this bug that takes another approach.