WIP - Fix save bread when ajax calls are not done
Every relationship set in BREAD made three ajax calls to collect table fields.
In a BREAD with many relationship and/or slow server this could result in some relationship not correctly populated when hitting Submit button.

This PR prepopulate selects with already saved values, since we have it, and leaves ajax call only when table is changed or modal is opened.

Fixes #3960

TODO: Fix select width in modal