Update multilanguage.md
In order to avoid headaches for others trying to implement Voyager's multilingual feature. I thought it would be positive to improve this part of the documentation. Some people had this problem in 2018 and today it is still happening in the current version (1.4).

I found this thread that helped me solve the problem in my own project:

https://github.com/the-control-group/voyager/issues/3172

I hope that you can accept the changes so that you can help others in the future. Awesome project, thanks!